refactor: add useArrangeViewKeyboardShortcut
