+++ app/src/components/ArrangeView/ArrangeEditor.tsx
@@ -1,8 +1,8 @@
 import styled from "@emotion/styled"
 import { FC } from "react"
 import { ArrangeViewScope } from "../../hooks/useArrangeView"
+import { useArrangeViewKeyboardShortcut } from "../../hooks/useArrangeViewKeyboardShortcut"
 import { ArrangeToolbar } from "../ArrangeToolbar/ArrangeToolbar"
-import { ArrangeViewKeyboardShortcut } from "../KeyboardShortcut/ArrangeViewKeyboardShortcut"
 import { ArrangeTransposeDialog } from "../TransposeDialog/ArrangeTransposeDialog"
 import { ArrangeVelocityDialog } from "../VelocityDialog/ArrangeVelocityDialog"
 import { ArrangeView } from "./ArrangeView"
@@ -13,13 +13,15 @@ const Container = styled.div`
   flex-direction: column;
   flex-grow: 1;
   position: relative;
+  outline: none;
 `
 
 export const ArrangeEditor: FC = () => {
+  const keyboardShortcutProps = useArrangeViewKeyboardShortcut()
+
   return (
     <ArrangeViewScope>
-      <Container>
-        <ArrangeViewKeyboardShortcut />
+      <Container {...keyboardShortcutProps} tabIndex={0}>
         <ArrangeToolbar />
         <ArrangeView />
       </Container>
